debug ui

+++ src/segment_intersection.c
@@ -3,7 +3,7 @@
 
 // From: https://en.wikipedia.org/wiki/Cohen%E2%80%93Sutherland_algorithm
 // slope = (y1 - y0) / (x1 - x0)
-// x = x0 + (1 / slope) * (ym - y0) where yw is ymin or ymax
+// slope = (b->y - a->y) / (b->x - a->x)
 // y = y0 + slope * (xm - x0), where xm is xmin or xmax
 
 float slope(t_vec2 *a, t_vec2 *b)
@@ -13,7 +13,8 @@ float slope(t_vec2 *a, t_vec2 *b)
 
 t_vec2 segment_intersect_x(t_vec2 *a, t_vec2 *b, int y)
 {
-    return ((t_vec2){a->x + (1 / slope(a, b)) * (y - a->y), y});
+    // x = x0 + (1 / slope) * (ym - y0) where yw is ymin or ymax
+    return ((t_vec2){(float)a->x + (1.0 / slope(a, b)) * (y - a->y), y});
 }
 
 t_vec2 segment_intersect_y(t_vec2 *a, t_vec2 *b, int x)
